Fix attribute byte output so that the last row is correct.
Attributes have an annoying quirk where there are 15 vertical blocks and 16 vertical blocks but those blocks are encoded in 2x2 groups. This means the bottom-most block row is encoded with the non-existent 16th row, which can just assumed to be 0. We punted on implementing it earlier, and never got around to fixing it.

----------------
real    0m0.344s
user    0m0.307s
sys     0m0.031s