Change palette add method so that it doesn’t remove bg_color.
Previously, adding colors to a palette would remove the bg_color because the best_palette guesser needed this, since it worked with color sets. This is not typical, however. Change the add method so that it replaces the first color to be the bg_color, and fix the guess_best_palette class so it handles this case itself.

----------------
real    0m0.343s
user    0m0.309s
sys     0m0.028s