Sprite mode, enabled with the -s flag.
Sprite mode is for images composed of sprites, not nametable. This has three effects. 1) Nametable and attribute components are not output, but spritelist component is. 2) The palette is stored at 0x10-0x1f instead of 0x00-0x0f. 3) The default order is set to 0x1000 instead of 0x0000, this can be overridden with the -r flag.

----------------
real    0m0.369s
user    0m0.287s
sys     0m0.045s