Fix bug where guessed palette might not contain the background color.
In cases where the -b flag is given, to set a background color, but the full palette is not explicitly provided, the guesser might build a palette set where the background color wasn't included. Then, when it would later get merged in, it would create a palette option with 5 colors.

----------------
real    0m0.386s
user    0m0.287s
sys     0m0.044s