Background color spec.
The background color flag -b, can be either a single hex value, or a pair separated by an equals sign. This pair contains the look which is used to process the pixel art image, then the fill which is output to the palette. More palette tests.

----------------
real    0m0.332s
user    0m0.278s
sys     0m0.041s