Support for Arduboy image format.
The Arduboy is an arduino based game system, that uses an image format that is similar to the NES. Images are stored as separate 1-bit planes, in a vertical-pixel orientation. Add flags that allow an easy way to generate this binary format.

----------------
real    0m0.309s
user    0m0.257s
sys     0m0.041s